+++ src/resource.h
@@ -3,7 +3,7 @@
 #define IDC_CLOSED_HAND 101
 #define IDC_OPEN_HAND   102
 
-#define VER_FILEVERSION                 2,1,2,0
-#define VER_FILEVERSION_STR             "2.1.2\0"
+#define VER_FILEVERSION                 2,1,3,0
+#define VER_FILEVERSION_STR             "2.1.3\0"
 #define VER_PRODUCTVERSION              VER_FILEVERSION
 #define VER_PRODUCTVERSION_STR          VER_FILEVERSION_STR